Restructure to avoid skipping blocks that don't commit properly
I noticed in testing that some blocks were getting skipped due to
errors that happened on committing. Rather than retrying the commit
of that block (which admittedly may lead to hanging) that block was
skipped over.

This is because when commit errors occurred, it called "continue",
which I originally meant to continue on the UPDATELOOP, which would
retry processing the same chainUpdate, but because that got moved
inside the pending batch loop, the "continue" moved to the next
pending batch and never retried the chainUpdate.

Another weird side effect of this was that if there were multiple
pending batches in a single chainUpdate, the megastatement was
appended to throughout the chainUpdate, but executed at the end of
each pendingBatch, which was certainly not the intended behavior,
though it may have functioned without any issues (though I suspect
that because we've been using websockets for testing, this never
arose because there were never multiple pending batches in a chain
update on websockets).

It's worth noting that these changes may end up putting us in infinite
loops where we cannot process a block and move on, but in general
(especially during testing) that's preferable to skipping a block
and not noticing if we weren't paying close attention to the logs.