Several changes to the project to facilitate flume light.
A heavy package has been added to enable http requests to a provided "flume-heavy" instance.

The api has been updated such that some of the methods can be routed to heavy if necessary. (This needs more work to be complete)

The consumer now has a process by which to check if an istance is heavy or light and direct the block consumption accordingly.

The config has been elaborated to include a heavy server string.