Replace coinbase index with coinbaseCompound
The accountBlocksMined query was fairly inefficient using the
coinbase index as it had to build a temp B-TREE to figure out
which blocks / transactions to return. Adding the block number
to the coinbaseCompound index doesn't have much size impact
or any significant write performance impact, but has substantial
returns on the query performance.