More selective index determination
When sqlite has to choose between address and topic0 indexes, in
some cases it can choose very badly. Typically address is the better
choice, but for some address / topic0 combinations, topic0 is a much
better choice.

Getting counts of records on an index tends to be very fast, so we can
query how many records are returned in just the `address` part of the
query and just the `topic0` part of the query, and start the larger
query with the smaller result set.

This does add a little bit of overhead, but it will only be done when
the index would be either topic0 or address, and avoiding really bad
index selection is probably worth the overhead.