Changes made to support appending bor logs to GetLogs call.
There is a hook in GetLogs which calls out to the polygon plugin. If present the plugin modifies the query string an params slice to enable the function to access bor logs and bring the data through. A migration has been added to the bor database which makes bor_logs a one to one map of event_logs by adding indexes.