A utility function has been added to getTransactionCount to check if tx is present locally.
The purpose here is to relinquish flume heavy from the burden of having to keep a mempool. This commit will serve as the MVP for functionality from an API
standpoint. There is further development to do to enable the heavy instances to drop their mempools and not sevice them in the future. Said work will be
appended to this branch but may need to be taken elsewhere depending on complexity.