Added uncertaunty check.
At this point the check will only work on flume instances with attached block databases. Also the check is rudementary in that it only scans for missing
blocks or block ranges. Thus we are assuming that if there are no blocks missing then the other data is present as well.